chore: forge fmt

+++ src/test/mocks/OperatorTableCalculatorMock.sol
@@ -9,7 +9,7 @@ contract OperatorTableCalculatorMock is IOperatorTableCalculator {
 
     mapping(bytes32 => bytes) internal _operatorTableBytes;
     mapping(bytes32 => address[]) internal _operators;
-    mapping(bytes32 => mapping(address => uint256)) internal _operatorWeights;
+    mapping(bytes32 => mapping(address => uint)) internal _operatorWeights;
 
     function calculateOperatorTableBytes(OperatorSet memory operatorSet) external view returns (bytes memory) {
         return _operatorTableBytes[operatorSet.key()];
@@ -19,25 +19,24 @@ contract OperatorTableCalculatorMock is IOperatorTableCalculator {
         _operatorTableBytes[operatorSet.key()] = operatorTableBytes;
     }
 
-    function getOperatorWeights(
-        OperatorSet calldata operatorSet
-    ) external view returns (address[] memory operators, uint256[][] memory weights) {
+    function getOperatorWeights(OperatorSet calldata operatorSet)
+        external
+        view
+        returns (address[] memory operators, uint[][] memory weights)
+    {
         bytes32 key = operatorSet.key();
         operators = _operators[key];
-        
-        weights = new uint256[][](operators.length);
-        for (uint256 i = 0; i < operators.length; i++) {
-            weights[i] = new uint256[](1);
+
+        weights = new uint[][](operators.length);
+        for (uint i = 0; i < operators.length; i++) {
+            weights[i] = new uint[](1);
             weights[i][0] = _operatorWeights[key][operators[i]];
         }
-        
+
         return (operators, weights);
     }
 
-    function getOperatorWeight(
-        OperatorSet calldata operatorSet,
-        address operator
-    ) external view returns (uint256 weight) {
+    function getOperatorWeight(OperatorSet calldata operatorSet, address operator) external view returns (uint weight) {
         return _operatorWeights[operatorSet.key()][operator];
     }
 
@@ -46,7 +45,7 @@ contract OperatorTableCalculatorMock is IOperatorTableCalculator {
         _operators[operatorSet.key()] = operators;
     }
 
-    function setOperatorWeight(OperatorSet memory operatorSet, address operator, uint256 weight) external {
+    function setOperatorWeight(OperatorSet memory operatorSet, address operator, uint weight) external {
         _operatorWeights[operatorSet.key()][operator] = weight;
     }
 }
